backend/requirements.txt
fastapi==0.115.8
email-validator==2.2.0
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.9
httpx==0.27.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.14.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
neo4j==5.24.0
cachetools==5.5.0
chromadb==0.5.5
rank-bm25==0.2.2
numpy==1.26.4
orjson==3.10.7
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2.post1
psycopg[binary]==3.2.6